### 2019-02-12 #1248 @whedon
[PRE REVIEW]: iced: fast and memory efficient normalization of contact maps
**Submitting author:** @NelleV (<a href="http://orcid.org/0000-0002-8748-6546">Nelle varoquaux</a>)
**Repository:** <a href="https://github.com/hiclib/iced" target ="_blank">https://github.com/hiclib/iced</a>
**Version:** v0.5.1
