### 2019-02-14 #1248 @NelleV
The labels from whedon are not quite accurate: this is primarly a Python library.
