### 2019-02-14 #1248 @pjotrp
I think they are ticked automatically. Fixed.
