### 2019-02-27 #1248 @NelleV
Hi. So sorry, I missed the first notification.
Do you need one or two reviewers? I can probably compile a list of people from the Hi-C community, but they might not be python developers.
