### 2019-03-01 #1286 @NelleV
The proofs look fine. I can tick everything, except Performance (I don't think it si applicaable) and Community guidelines (I'm not sure what is expected there--I've got a brief statement in the first page of the documentation that I am happy to extend if it is not detailed enough: http://projects.cbio.mines-paristech.fr/iced/index.html)
