### 2019-03-15 hiclib/iced#61 @robinweide
plain-text LICENSE file
Could you rename/copy the COPYING-file to LICENSE? This makes it easier to find.
