### 2019-03-15 hiclib/iced#62 @robinweide
community guidelines links to pastis-issue tracker
In the [docs](http://projects.cbio.mines-paristech.fr/iced/index.html) you link to this issue-tracker: https://github.com/hiclib/pastis.

I think it would be more logical to link to the issue-tracker of iced.
