### 2019-03-15 #1286 @robinweide
Hi @NelleV, @pjotrp,

I think it lookes quite good. The documentation is expansive and all code is clear and commented where applicable.

Minor ssues I have:

- [LICENSE-file missing](https://github.com/hiclib/iced/issues/61)
-  @NelleV you can tick off **performance** if you cite Servant et al. (2015) for the iced-benchmarks
- [community guidelines links to pastis-issue tracker](https://github.com/hiclib/iced/issues/62)
- I cannot find the DOI's in the references.
