### 2019-03-21 81982bf
FIX sparse matrix is more strict on types
