### 2019-03-21 fa741d2
MAINT COPYING -> LICENSE
closes #61
