### 2019-03-21 #1286 @NelleV
Thanks @robinweide for the review!

I've fixed all items be the number 3 (community guidelines). Unless I'm looking at the wrong link in the documentation, it's already pointing to the issue tracker.
