### 2019-03-22 hiclib/iced#62 @robinweide
In the docs under "CONTACTS":
_If you have any questions or suggestions, please email nelle dot varoquaux at ensmp dot fr, or open a ticket on Github_

The link under github goes to https://github.com/hiclib/pastis/issues
