### 2019-03-22 b05c6f0
FIX right link in issue tracker
closes #62
