### 2019-03-22 hiclib/iced#62 @NelleV
Oh wow. Thanks!

On Fri, 22 Mar 2019 at 03:48, Robin van der Weide <notifications@github.com>
wrote:

> In the docs under "CONTACTS":
> *If you have any questions or suggestions, please email nelle dot
> varoquaux at ensmp dot fr, or open a ticket on Github*
>
> The link under github goes to https://github.com/hiclib/pastis/issues
>
