### 2019-03-25 #1286 @NelleV
For information, point n3 is now also fixed. The link was pointing to the wrong bug tracker…

I hope this is now good to go. Let me know if I need to do anything else!
