### 2019-03-26 #1286 @robinweide
> @robinweide are you happy with the fixes?

@pjotrp yes. However, since I'm already a user of HiC-pro -and thus iced- for a long time, I do strongly urge you to find a second reviewer. He/she could verify if all (installation/basic operation) is clear for a novice (Hi-C)user.
