### 2019-03-27 #1286 @pjotrp
Dear @NelleV, do you mind starting the paper with a short laypersons introduction of HiC and what your code does? Just paragraph will do to help the casual reader.
