### 2019-03-27 #1286 @NelleV
Yes, no problem. Can I consider the article accepted with minor revisions? I have a job talk tomorrow.
