+++ doc/papers/paper.md
@@ -27,9 +27,19 @@ bibliography: paper.bib
 
 # Summary
 
+The three-dimensional structure of the genome is thought to play an important
+role in many biological processes, such as gene regulation and replication.
 Recent technological advances allow the measurement, in a single Hi-C
 experiment, of the frequencies of physical contacts among pairs of genomic
-loci at a genome-wide scale.
+loci at a genome-wide scale. The Hi-C protocol results in a noisy and indirect
+measurement of the 3D structure of the genome, yielding a symmetric matrix
+where each row and each column correspond to a genomic window, and each entry
+to the number of times those windows have been seen interacting with one
+another.  As any genomics experiments, the resulting matrix contains unwanted
+variations depending on the GC-content, mappability, and the details of the
+protocol used.
+
+![](figures/counts_pfalc.png)
 
 [**Iced**](https://github.com/hiclib/iced) implements fast and memory
 efficient normalization methods, such the ICE normalization strategy or the
