### 2019-03-27 #1286 @NelleV
Can I re-trigger whedon's pdf generation?
