### 2019-03-27 #1286 @danielskatz
just put ```@whedon generate pdf``` in as a new comment
