### 2019-03-27 #1286 @pjotrp
> Yes, no problem. Can I consider the article accepted with minor revisions? I have a job talk tomorrow.

Yes
