### 2019-03-31 #1286 @pjotrp
Oh, I see it now. @NelleV you need to remove the references in the bibtex file that you do not use in the paper. This is a recent check added to @whedon
