### 2019-04-02 #1286 @NelleV
I've deposited the code in OSF already. Is that a suitable DOI issuer?
The DOI is 10.17605/OSF.IO/N6WHX
