### 2019-04-02 1ba02c7
Updating DOI for bibliography
