+++ doc/papers/paper.bib
@@ -5,7 +5,8 @@ and R. Koszul and J. Mozziconacci},
   Journal                  = {BMC Genomics},
   Year                     = {2012},
   Pages                    = {436},
-  Volume                   = {13}
+  Volume                   = {13},
+  doi="10.1186/1471-2164-13-436"
 }
 
 
@@ -18,7 +19,8 @@ and R. Koszul and J. Mozziconacci},
   Journal                  = {Nature Methods},
   Year                     = {2012},
   Pages                    = {999--1003},
-  Volume                   = {9}
+  Volume                   = {9},
+  doi = "10.1038/nmeth.2148"
 }
 
 @Article{li:hi-corrector,
@@ -30,7 +32,8 @@ normalizing large-scale {H}i-{C} data}",
    Volume="31",
    Number="6",
    Pages="960--962",
-   Month="Mar"
+   Month="Mar",
+   doi="10.1093/bioinformatics/btu747"
 }
 
 @Article{varoquaux:iced_osf,
@@ -48,4 +51,5 @@ processing}",
    Journal="Genome Biol.",
    Year="2015",
    Volume="16",
+   doi="10.1186/s13059-015-0831-x"
 }
