### 2019-04-02 #1286 @NelleV
I've added the DOI for all of the references.
