### 2019-04-03 #1286 @pjotrp
Whedon accepts the DOI. Maybe the EIC team wants to have a look.

ping eic @openjournals/joss-eics
