### 2019-04-03 #1286 @kyleniemeyer
Hi @NelleV, just a few final suggestions/corrections for the article:

- I think "As any genomics experiments" is missing a word.
- You don't need to cite the software in the paper, since there is already a link to the archive, and this paper serves as a reference for the software. (If you really want to keep it, it does need some fixing—th names aren't showing up correct, and it should have 2019 for the year.)
- The first sentence is great, but the second sentence mentions Hi-C which isn't explained—for nonexperts it might be nice to give a few words explaining what this is, since the rest of this paragraph is generally understandable.

Otherwise, looks great!
