### 2019-04-03 dd7763c
Small fixes on the paper
