### 2019-04-03 #1286 @kyleniemeyer
@NelleV sorry to come back with another question, but I'm a little confused about the OSF DOI—from what I can tell, OSF is not actually archiving the accepted version of the software, but really linking to the GitHub repo, which can change. For our purposes, we really want an archived snapshot of the accepted version that could be downloaded, and usually people use Zenodo or Figshare for this purpose.

(I don't think @whedon does any checking of the archive DOI, other than that it's legit. @arfon perhaps you could comment?)
