+++ doc/papers/paper.bib
@@ -39,7 +39,7 @@ normalizing large-scale {H}i-{C} data}",
 @Article{varoquaux:iced_osf,
     Author="Varoquaux, N, and Servant, N",
     Title="iced: fast and memory efficient normalization of contact maps",
-    doi="10.17605/OSF.IO/N6WHX"
+    doi="10.5281/zenodo.2622857"
     }
 
 @article{servant:hicpro,
