### 2019-04-03 #1286 @arfon
> (I don't think @whedon does any checking of the archive DOI, other than that it's legit. @arfon perhaps you could comment?)

That's correct.

> I think everything is fine now, but the archive link on the article proof is not working.

This is only 'linked up' in the final proofs.
