### 2019-04-03 #1286 @kyleniemeyer
Congrats @NelleV on your article published in JOSS!  Thanks @robinweide for reviewing and @pjotrp for editing.
